requirements.txt
confluent-kafka~=2.5.3
faust-streaming~=0.11.3
websockets~=13.0.1
pyspark~=3.5.0
numpy~=1.26.0
pandas~=2.1.0
dbldatagen~=0.4.0.post1
apache-beam[interactive,dataframe]~=2.58.0
redis~=5.0.8
kombu~=5.3.4
httpx~=0.27.2 
